Benchmarking (#12)
* Adding test data

* Rarefaction, is faster than that of vegan!

* Removing pprof from main

* if you are less than or equal to the sampled size, we do not need to rarefy

* Modifications to the new subsampling function.

* Trimming fat, and fixing test

* Rarefaction fix (#11)

* Adding rarefaction modifications. These arent for speed, but for testing

* Lets start from the beginning...It takes 11 seconds on average to rarefy a community matrix with 3261 cols and 28 rows.

* Adding community matrix class

* Creating functionality for the community matrix class.

* Todo comments

* May have found a method to properly rarefy

* Removing comments

* Squashed commit of the following:

commit 20181ac02bce438ced6b99107bfdb11578a4e551
Author: YoungKrug <gregjohn456@gmail.com>
Date:   Sat Mar 1 22:04:50 2025 -0500

    Trying a different approach

* Adding function for calculating diversity using a community matrix object

* It will now skip a row if you try to rarefy to an impossible threshold.

* Fixing the package structure and ensuring all the test work. Getting ready to push back to dev

* Fixing documentation files and test to reflect recent changes.

* Removing test files

* Removed cpp build warnings

* Added massdatabase dependency

+++ src/GNPSScoringDynamicPriorityQueue.cpp
@@ -54,7 +54,7 @@ std::unordered_map<int, std::unordered_set<int>> GNPSScoringDynamicPriorityQueue
     std::unordered_map<int, std::unordered_set<int>> indexMap;
     totalSizeOfIndexMap = 0;
     // Iterate over
-    for(int i = 0; i < mzVector.size(); i++)
+    for(size_t i = 0; i < mzVector.size(); i++)
     {
         const double highBounds = mzVector[i] + adjTolerance - shift;
         const double lowBounds = mzVector[i]  - adjTolerance - shift;
@@ -106,7 +106,7 @@ double GNPSScoringDynamicPriorityQueue::ScoreMatches(
         usedPeakTwo.insert(indexTwo);
         totalScore += score;
         numberOfPeakMatches ++;
-        if(numberOfPeakMatches >= countOfSpectraOne)
+        if(numberOfPeakMatches >= static_cast<int>(countOfSpectraOne))
             break;
         std::pop_heap(queue.begin(), queue.end(), CompareScores());
     }
